requirements.txt
# Runtime dependencies
pyserial